V3.1.2
* Allow for decimals to be entered
* Fixed bug where backspacing on input dialogues for specific browsers
would not work

+++ javascripts/init.js
@@ -201,7 +201,7 @@ $( document ).ready(function() {
     });
 
     if($("#numberInput").val().length > 0){
-      NUMBER_RETURNED= parseInt($("#numberInput").val());
+      NUMBER_RETURNED= parseFloat($("#numberInput").val());
     }
     $("#numberInput").val(getCookie("numberInput"));
     $("#lower-margin-threshold").val(getCookie("lower-margin-threshold"));
@@ -213,7 +213,7 @@ $( document ).ready(function() {
 });
 
 function updateNumber(){
-    NUMBER_RETURNED = parseInt($("#numberInput").val());
+    NUMBER_RETURNED = parseFloat($("#numberInput").val());
     if(NUMBER_RETURNED > 10 || NUMBER_RETURNED < 1){
         NUMBER_RETURNED = 1;
         $("#numberInput").val("1");
@@ -329,14 +329,14 @@ function init(){
         }
 
         if($("#lower-margin-threshold").val().length > 0 && !isNaN($("#lower-margin-threshold").val())){
-            threshold_margin_lower = parseInt($("#lower-margin-threshold").val());
+            threshold_margin_lower = parseFloat($("#lower-margin-threshold").val());
             setCookie("lower-margin-threshold",threshold_margin_lower,7);
         }else{
             setCookie("lower-margin-threshold","");
         }
 
         if($("#upper-margin-threshold").val().length > 0 && !isNaN($("#upper-margin-threshold").val())){
-            threshold_margin_upper = parseInt($("#upper-margin-threshold").val());
+            threshold_margin_upper = parseFloat($("#upper-margin-threshold").val());
             setCookie("upper-margin-threshold",threshold_margin_upper,7);
         }else{
             setCookie("upper-margin-threshold","");
@@ -362,25 +362,25 @@ function init(){
       }
 
       if($("#profit-threshold").val().length > 0 && !isNaN($("#profit-threshold").val())){
-          threshold_profit = parseInt($("#profit-threshold").val());
+          threshold_profit = parseFloat($("#profit-threshold").val());
           setCookie("profit-threshold",threshold_profit,7);
       }else{
           setCookie("profit-threshold","");
       }
       if($("#roi-threshold").val().length > 0 && !isNaN($("#roi-threshold").val())){
-          threshold_roi = parseInt($("#roi-threshold").val());
+          threshold_roi = parseFloat($("#roi-threshold").val());
           setCookie("roi-threshold",threshold_roi,7);
       }else{
           setCookie("roi-threshold","");
       }
       if($("#buy-threshold").val().length > 0 && !isNaN($("#buy-threshold").val())){
-          threshold_cost = parseInt($("#buy-threshold").val());
+          threshold_cost = parseFloat($("#buy-threshold").val());
           setCookie("buy-threshold",threshold_cost,7);
       }else{
           setCookie("buy-threshold","");
       }
       if($("#weight-threshold").val().length > 0 && !isNaN($("#weight-threshold").val())){
-          threshold_weight = parseInt($("#weight-threshold").val());
+          threshold_weight = parseFloat($("#weight-threshold").val());
           setCookie("weight-threshold",threshold_weight,7);
       }else{
           setCookie("weight-threshold","");
